syncmanagerapi/pyproject.toml
[tool.poetry]
name = "syncmanagerapi"
version = "0.1.0"  # Replace with the actual version
description = "Provides Server for managing multiple synchronizations via unison, git, ..."
authors = ["Gerd Friemel <your.email@example.com>"]  # Replace with actual authors
license = "MIT"  # Replace with the correct license if applicable
readme = "README.md"
repository = "https://github.com/Grid-LTS/syncmanager"
homepage = "https://github.com/Grid-LTS/syncmanager"
packages = [{ include = "syncmanagerapi" }]


[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
connexion = {extras = ["flask", "swagger-ui", "uvicorn"], version = "^3.2.0"}
dotenv = "^0.9.9"
gitpython = "^3.1.44"
flask-basicauth = "^0.2.0"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.2"
gunicorn = "^23.0.0"
swagger-ui-bundle = "^1.1.0"
mysqlclient = "^2.2.7"
jinja2 = "^3.1.6"
uvicorn = "^0.34.2"
alembic = "^1.15.2"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "syncmanagerapi"
version = "0.1.0"  # Replace with the actual version
description = "Provides Server for managing multiple synchronizations via unison, git, ..."


[tool.setuptools.packages.find]
where = ["syncmanagerapi"]

